requirements.txt
streamlit
transformers
PyPDF2
python-docx
fpdf
torch
google-generativeai
pdfplumber
gradio==4.28.0
rouge-score
transformers
torch
nltk
scikit-learn
